+++ driver/sql/postgres/internal/pgerror/doc.go
@@ -1,0 +1,2 @@
+// Package pgerror provides utilities for working with PostgreSQL errors.
+package pgerror