+++ driver/sql/postgres/pgkv/schema.go
@@ -5,25 +5,27 @@ import (
 	"database/sql"
 
 	_ "embed"
+
+	"github.com/dogmatiq/persistencekit/driver/sql/postgres/internal/pgerror"
 )
 
 //go:embed schema.sql
 var schema string
 
-// CreateSchema creates the PostgreSQL schema elements required by [Store].
-func CreateSchema(
+// createSchema creates the PostgreSQL schema elements required by [Store].
+func createSchema(
 	ctx context.Context,
 	db *sql.DB,
 ) error {
-	tx, err := db.Begin()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
-
-	if _, err := tx.ExecContext(ctx, schema); err != nil {
-		return err
-	}
-
-	return tx.Commit()
+	return pgerror.Retry(
+		ctx,
+		db,
+		func(tx *sql.Tx) error {
+			_, err := tx.ExecContext(ctx, schema)
+			return err
+		},
+		// Even though we use IF NOT EXISTS in the DDL, we still need to handle
+		// conflicts due to a data race bug in PostgreSQL.
+		pgerror.CodeUniqueViolation,
+	)
 }